Issue application sends AoS Pack and notification
