+++ src/test/java/uk/gov/hmcts/divorce/caseworker/service/CaseworkerIssueApplicationServiceTest.java
@@ -1,45 +1,36 @@
 package uk.gov.hmcts.divorce.caseworker.service;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
-import uk.gov.hmcts.ccd.sdk.api.callback.AboutToStartOrSubmitResponse;
-import uk.gov.hmcts.divorce.caseworker.service.notification.NoticeOfProceedingsNotification;
-import uk.gov.hmcts.divorce.caseworker.service.notification.PersonalServiceNotification;
-import uk.gov.hmcts.divorce.caseworker.service.print.AosPackPrinter;
 import uk.gov.hmcts.divorce.caseworker.service.updater.MiniApplication;
 import uk.gov.hmcts.divorce.caseworker.service.updater.RespondentSolicitorAosInvitation;
-import uk.gov.hmcts.divorce.divorcecase.model.AcknowledgementOfService;
+import uk.gov.hmcts.divorce.caseworker.service.updater.SendAosNotifications;
+import uk.gov.hmcts.divorce.caseworker.service.updater.SendAosPack;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
 import uk.gov.hmcts.divorce.divorcecase.model.Solicitor;
-import uk.gov.hmcts.divorce.divorcecase.model.State;
 import uk.gov.hmcts.divorce.divorcecase.updater.CaseDataContext;
 import uk.gov.hmcts.divorce.divorcecase.updater.CaseDataUpdater;
 import uk.gov.hmcts.divorce.divorcecase.updater.CaseDataUpdaterChain;
 import uk.gov.hmcts.divorce.divorcecase.updater.CaseDataUpdaterChainFactory;
 
 import java.time.Clock;
-import java.time.Instant;
 import java.time.LocalDate;
-import java.time.ZoneId;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
-import static org.springframework.test.util.ReflectionTestUtils.setField;
 import static uk.gov.hmcts.ccd.sdk.type.YesOrNo.NO;
 import static uk.gov.hmcts.ccd.sdk.type.YesOrNo.YES;
-import static uk.gov.hmcts.divorce.divorcecase.model.ServiceMethod.COURT_SERVICE;
-import static uk.gov.hmcts.divorce.divorcecase.model.ServiceMethod.PERSONAL_SERVICE;
-import static uk.gov.hmcts.divorce.divorcecase.model.State.AwaitingAos;
+import static uk.gov.hmcts.divorce.testutil.ClockTestUtil.NOW;
+import static uk.gov.hmcts.divorce.testutil.ClockTestUtil.ZONE_ID;
+import static uk.gov.hmcts.divorce.testutil.ClockTestUtil.getExpectedLocalDate;
+import static uk.gov.hmcts.divorce.testutil.ClockTestUtil.setMockClock;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_AUTHORIZATION_TOKEN;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_CASE_ID;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_ORG_NAME;
@@ -47,24 +38,9 @@ import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_SOLICITOR_EMAIL;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_SOLICITOR_EMAIL;
 import static uk.gov.hmcts.divorce.testutil.TestDataHelper.LOCAL_DATE;
 import static uk.gov.hmcts.divorce.testutil.TestDataHelper.caseData;
-import static uk.gov.hmcts.divorce.testutil.TestDataHelper.respondent;
-import static uk.gov.hmcts.divorce.testutil.TestDataHelper.respondentWithDigitalSolicitor;
 
 @ExtendWith(MockitoExtension.class)
 class CaseworkerIssueApplicationServiceTest {
-
-    private static final Instant NOW = Instant.now();
-    private static final ZoneId ZONE_ID = ZoneId.systemDefault();
-    private static final long DUE_DATE_OFFSET_DAYS = 30L;
-
-    @Mock
-    private PersonalServiceNotification personalServiceNotification;
-
-    @Mock
-    private NoticeOfProceedingsNotification noticeOfProceedingsNotification;
-
-    @Mock
-    private AosPackPrinter aosPackPrinter;
 
     @Mock
     private MiniApplication miniApplication;
@@ -76,15 +52,16 @@ class CaseworkerIssueApplicationServiceTest {
     private CaseDataUpdaterChainFactory caseDataUpdaterChainFactory;
 
     @Mock
+    private SendAosPack sendAosPack;
+
+    @Mock
+    private SendAosNotifications sendAosNotifications;
+
+    @Mock
     private Clock clock;
 
     @InjectMocks
     private IssueApplicationService issueApplicationService;
-
-    @BeforeEach
-    void setPageSize() {
-        setField(issueApplicationService, "dueDateOffsetDays", DUE_DATE_OFFSET_DAYS);
-    }
 
     @Test
     void shouldGenerateMiniApplicationAndRespondentAosAndSetIssueDateWhenRespondentIsSolicitorRepresented() {
@@ -102,7 +79,11 @@ class CaseworkerIssueApplicationServiceTest {
 
         final CaseDataUpdaterChain caseDataUpdaterChain = mock(CaseDataUpdaterChain.class);
 
-        final List<CaseDataUpdater> caseDataUpdaters = List.of(miniApplication, respondentSolicitorAosInvitation);
+        final List<CaseDataUpdater> caseDataUpdaters = List.of(
+            miniApplication,
+            respondentSolicitorAosInvitation,
+            sendAosPack,
+            sendAosNotifications);
 
         final CaseDataContext caseDataContext = CaseDataContext.builder()
             .caseData(caseData)
@@ -111,7 +92,7 @@ class CaseworkerIssueApplicationServiceTest {
             .userAuthToken(TEST_AUTHORIZATION_TOKEN)
             .build();
 
-        setClock();
+        setMockClock(clock);
 
         when(caseDataUpdaterChainFactory.createWith(caseDataUpdaters)).thenReturn(caseDataUpdaterChain);
         when(caseDataUpdaterChain.processNext(caseDataContext)).thenReturn(caseDataContext);
@@ -123,10 +104,8 @@ class CaseworkerIssueApplicationServiceTest {
             TEST_AUTHORIZATION_TOKEN
         );
 
-        final var expectedDateTime = LocalDate.ofInstant(NOW, ZONE_ID);
-
         var expectedCaseData = caseData();
-        expectedCaseData.getApplication().setIssueDate(expectedDateTime);
+        expectedCaseData.getApplication().setIssueDate(getExpectedLocalDate());
         expectedCaseData.getApplicant2().setSolicitorRepresented(YES);
         expectedCaseData.getApplicant2().setSolicitor(solicitor);
 
@@ -146,7 +125,10 @@ class CaseworkerIssueApplicationServiceTest {
 
         final CaseDataUpdaterChain caseDataUpdaterChain = mock(CaseDataUpdaterChain.class);
 
-        final List<CaseDataUpdater> caseDataUpdaters = List.of(miniApplication);
+        final List<CaseDataUpdater> caseDataUpdaters = List.of(
+            miniApplication,
+            sendAosPack,
+            sendAosNotifications);
 
         final CaseDataContext caseDataContext = CaseDataContext.builder()
             .caseData(caseData)
@@ -155,7 +137,7 @@ class CaseworkerIssueApplicationServiceTest {
             .userAuthToken(TEST_AUTHORIZATION_TOKEN)
             .build();
 
-        setClock();
+        setMockClock(clock);
 
         when(caseDataUpdaterChainFactory.createWith(caseDataUpdaters)).thenReturn(caseDataUpdaterChain);
         when(caseDataUpdaterChain.processNext(caseDataContext)).thenReturn(caseDataContext);
@@ -179,124 +161,5 @@ class CaseworkerIssueApplicationServiceTest {
         verify(caseDataUpdaterChain).processNext(caseDataContext);
 
         verifyNoMoreInteractions(caseDataUpdaterChainFactory, caseDataUpdaterChain);
-    }
-
-    @Test
-    void shouldNotPrintAosIfApplicationIsPersonalServiceMethodWhenAboutToSubmit() {
-
-        final var caseData = caseData();
-        caseData.getApplication().setSolServiceMethod(PERSONAL_SERVICE);
-        final var caseDetails = CaseDetails.<CaseData, State>builder()
-            .id(TEST_CASE_ID)
-            .state(AwaitingAos)
-            .data(caseData)
-            .build();
-
-        final AboutToStartOrSubmitResponse<CaseData, State> response = issueApplicationService.sendAosPack(caseDetails, null);
-
-        assertThat(response.getData().getDueDate()).isNull();
-        assertThat(response.getData().getAcknowledgementOfService())
-            .extracting(
-                AcknowledgementOfService::getDigitalNoticeOfProceedings,
-                AcknowledgementOfService::getNoticeOfProceedingsEmail,
-                AcknowledgementOfService::getNoticeOfProceedingsSolicitorFirm)
-            .contains(null, null, null);
-        verifyNoInteractions(aosPackPrinter, clock);
-    }
-
-    @Test
-    void shouldSendPersonalServiceNotificationIfPersonalServiceApplication() {
-
-        final var caseData = caseData();
-        caseData.getApplication().setSolServiceMethod(PERSONAL_SERVICE);
-        final var caseDetails = CaseDetails.<CaseData, State>builder()
-            .id(TEST_CASE_ID)
-            .state(AwaitingAos)
-            .data(caseData)
-            .build();
-
-        issueApplicationService.sendAosPack(caseDetails, null);
-
-        verify(personalServiceNotification).send(caseData, TEST_CASE_ID);
-        verifyNoInteractions(noticeOfProceedingsNotification);
-    }
-
-    @Test
-    void shouldPrintAosAndSetDueDateIfNotPersonalServiceAndRespondentIsNotRepresented() {
-
-        final var caseData = caseData();
-        caseData.getApplication().setSolServiceMethod(COURT_SERVICE);
-        caseData.setApplicant2(respondent());
-        final var caseDetails = CaseDetails.<CaseData, State>builder()
-            .id(TEST_CASE_ID)
-            .state(AwaitingAos)
-            .data(caseData)
-            .build();
-
-        setClock();
-        final LocalDate expectedDueDate = LocalDate.ofInstant(NOW, ZONE_ID).plusDays(DUE_DATE_OFFSET_DAYS);
-
-        final AboutToStartOrSubmitResponse<CaseData, State> response = issueApplicationService.sendAosPack(caseDetails, null);
-
-        assertThat(response.getData().getDueDate()).isEqualTo(expectedDueDate);
-        assertThat(response.getData().getAcknowledgementOfService())
-            .extracting(
-                AcknowledgementOfService::getDigitalNoticeOfProceedings,
-                AcknowledgementOfService::getNoticeOfProceedingsEmail,
-                AcknowledgementOfService::getNoticeOfProceedingsSolicitorFirm)
-            .contains(null, null, null);
-        verify(aosPackPrinter).print(caseData, TEST_CASE_ID);
-    }
-
-    @Test
-    void shouldPrintAosAndUpdateCaseDataIfNotPersonalServiceAndRespondentIsRepresented() {
-
-        final var caseData = caseData();
-        caseData.getApplication().setSolServiceMethod(COURT_SERVICE);
-        caseData.setApplicant2(respondentWithDigitalSolicitor());
-        final var caseDetails = CaseDetails.<CaseData, State>builder()
-            .id(TEST_CASE_ID)
-            .state(AwaitingAos)
-            .data(caseData)
-            .build();
-
-        setClock();
-        final LocalDate expectedDueDate = LocalDate.ofInstant(NOW, ZONE_ID).plusDays(DUE_DATE_OFFSET_DAYS);
-
-        final AboutToStartOrSubmitResponse<CaseData, State> response = issueApplicationService.sendAosPack(caseDetails, null);
-
-        assertThat(response.getData().getDueDate()).isEqualTo(expectedDueDate);
-        assertThat(response.getData().getAcknowledgementOfService())
-            .extracting(
-                AcknowledgementOfService::getDigitalNoticeOfProceedings,
-                AcknowledgementOfService::getNoticeOfProceedingsEmail,
-                AcknowledgementOfService::getNoticeOfProceedingsSolicitorFirm)
-            .contains(YES, TEST_SOLICITOR_EMAIL, TEST_ORG_NAME);
-        verify(aosPackPrinter).print(caseData, TEST_CASE_ID);
-    }
-
-    @Test
-    void shouldSendNoticeOfProceedingsIfNotPersonalServiceApplication() {
-
-        final var caseData = caseData();
-        caseData.getApplication().setSolServiceMethod(COURT_SERVICE);
-        caseData.setApplicant2(respondent());
-        final var caseDetails = CaseDetails.<CaseData, State>builder()
-            .id(TEST_CASE_ID)
-            .state(AwaitingAos)
-            .data(caseData)
-            .build();
-
-        setClock();
-
-        issueApplicationService.sendAosPack(caseDetails, null);
-
-        verify(noticeOfProceedingsNotification).send(caseData, TEST_CASE_ID);
-        verifyNoInteractions(personalServiceNotification);
-    }
-
-    private void setClock() {
-        when(clock.instant()).thenReturn(NOW);
-        when(clock.getZone()).thenReturn(ZONE_ID);
     }
 }
