+++ src/test/java/uk/gov/hmcts/divorce/caseworker/service/updater/RespondentSolicitorAosInvitationTest.java
@@ -9,7 +9,6 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.divorce.divorcecase.model.Applicant;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseInvite;
-import uk.gov.hmcts.divorce.divorcecase.updater.CaseDataContext;
 import uk.gov.hmcts.divorce.divorcecase.updater.CaseDataUpdaterChain;
 import uk.gov.hmcts.divorce.divorcecase.util.AccessCodeGenerator;
 import uk.gov.hmcts.divorce.document.CaseDataDocumentService;
@@ -31,6 +30,7 @@ import static uk.gov.hmcts.divorce.testutil.TestConstants.ACCESS_CODE;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_AUTHORIZATION_TOKEN;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_CASE_ID;
 import static uk.gov.hmcts.divorce.testutil.TestDataHelper.LOCAL_DATE;
+import static uk.gov.hmcts.divorce.testutil.UpdaterTestUtil.caseDataContext;
 
 @ExtendWith(MockitoExtension.class)
 public class RespondentSolicitorAosInvitationTest {
@@ -73,7 +73,7 @@ public class RespondentSolicitorAosInvitationTest {
                 templateContentSupplier,
                 TEST_CASE_ID,
                 TEST_AUTHORIZATION_TOKEN,
-                    RESP_SOLICITOR_AOS_INVITATION,
+                RESP_SOLICITOR_AOS_INVITATION,
                 RESP_AOS_INVITATION_DOCUMENT_NAME,
                 ENGLISH))
             .thenReturn(caseData);
@@ -85,15 +85,5 @@ public class RespondentSolicitorAosInvitationTest {
         assertThat(result.getCaseData()).isEqualTo(caseData);
 
         classMock.close();
-    }
-
-    private CaseDataContext caseDataContext(CaseData caseData) {
-        return CaseDataContext
-            .builder()
-            .caseData(caseData)
-            .caseId(TEST_CASE_ID)
-            .createdDate(LOCAL_DATE)
-            .userAuthToken(TEST_AUTHORIZATION_TOKEN)
-            .build();
     }
 }
