+++ src/test/java/uk/gov/hmcts/divorce/testutil/UpdaterTestUtil.java
@@ -1,0 +1,24 @@
+package uk.gov.hmcts.divorce.testutil;
+
+import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
+import uk.gov.hmcts.divorce.divorcecase.updater.CaseDataContext;
+
+import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_AUTHORIZATION_TOKEN;
+import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_CASE_ID;
+import static uk.gov.hmcts.divorce.testutil.TestDataHelper.LOCAL_DATE;
+
+public final class UpdaterTestUtil {
+
+    private UpdaterTestUtil() {
+    }
+
+    public static CaseDataContext caseDataContext(final CaseData caseData) {
+        return CaseDataContext
+            .builder()
+            .caseData(caseData)
+            .caseId(TEST_CASE_ID)
+            .createdDate(LOCAL_DATE)
+            .userAuthToken(TEST_AUTHORIZATION_TOKEN)
+            .build();
+    }
+}
